+++ cocos/scripting/lua-bindings/manual/platform/android/jni/Cocos2dxLuaJavaBridge.h
