+++ templates/cocos2dx_files.json
@@ -5517,8 +5517,8 @@
         "cocos/scripting/lua-bindings/manual/physics3d/lua_cocos2dx_physics3d_manual.h", 
         "cocos/scripting/lua-bindings/manual/platform/android/CCLuaJavaBridge.cpp", 
         "cocos/scripting/lua-bindings/manual/platform/android/CCLuaJavaBridge.h", 
-        "cocos/scripting/lua-bindings/manual/platform/android/jni/Java_org_cocos2dx_lib_Cocos2dxLuaJavaBridge.cpp", 
-        "cocos/scripting/lua-bindings/manual/platform/android/jni/Java_org_cocos2dx_lib_Cocos2dxLuaJavaBridge.h", 
+        "cocos/scripting/lua-bindings/manual/platform/android/jni/Cocos2dxLuaJavaBridge.cpp", 
+        "cocos/scripting/lua-bindings/manual/platform/android/jni/Cocos2dxLuaJavaBridge.h", 
         "cocos/scripting/lua-bindings/manual/platform/ios/CCLuaObjcBridge.h", 
         "cocos/scripting/lua-bindings/manual/platform/ios/CCLuaObjcBridge.mm", 
         "cocos/scripting/lua-bindings/manual/spine/LuaSkeletonAnimation.cpp",
